% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_metadata_generation.R
\name{variable_metadata_generation}
\alias{variable_metadata_generation}
\title{variable_metadata_generation}
\usage{
variable_metadata_generation(path_to_excel_directory,
  path_to_stata_directory, missing_conditions_file,
  path_to_output_directory, variables_no_distribution)
}
\description{
variable_metadata_generation
}
